packages/gkg-evals/pipeline/pyproject.toml
[project]
name = "gkg-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chardet>=5.2.0",
    "dataclasses-json>=0.6.7",
    "datasets>=4.0.0",
    "docker>=7.1.0",
    "ghapi>=1.0.6",
    "gitpython>=3.1.45",
    "httpx>=0.28.1",
    "modal>=1.1.4",
    "orjson==3.10.0",
    "pre-commit>=4.3.0",
    "pydantic>=2.11.7",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "swe-rex>=1.4.0",
    "tenacity>=9.1.2",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
    "unidiff>=0.7.5",
]

